Extract PetPicker from FeedPetDialog
